app/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
pymongo
redis
passlib[bcrypt]
python-dotenv
